Moved integration test class to a @require PHP >= 8.0 test case
